Removing Dead Links and Fixing Typo (#1137)
Been ramping up on GPUD and was noticing some dead links (and a small
typo) throughout the `xid.go` file.

Seems Nvidia moved around some links and restructured the doc that was
being linked to originally.

Let me know if this type of PR is allowed, just getting hands on with
GPUD so thought I'd start light.